Refactor reporting task so that methods can be used for the annual usage page. Still a work in progress, tests are coming.
